requirements.txt
openai
pyaudio
wave
pymongo[srv]
python-dateutil
